requirements.txt
groq
crewai
langchain-groq
streamlit
python-dotenv
pydantic
